Add query to Home
