+++ apps/consultation-portal/specs/index.spec.tsx
@@ -5,7 +5,7 @@ import Index from '../pages/index'
 
 describe('Index', () => {
   it('should render successfully', () => {
-    const { baseElement } = render(<Index />)
+    const { baseElement } = render(<Index cases={null} />)
     expect(baseElement).toBeTruthy()
   })
 })
